abcdxyznmadau/requirement.txt
transformers>=4.30.0
torch>=1.13.0
Pillow